profile/hrp/central/src/androidTest/java/org/im97mori/ble/profile/hrp/central/db/HeartRateProfileBondedDatabaseHelperTest.java
package org.im97mori.ble.profile.hrp.central.db;

import androidx.test.core.app.ApplicationProvider;

import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

public class HeartRateProfileBondedDatabaseHelperTest {

    @Test
    public void test_constructor_00001() {
        assertNotNull(HeartRateProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_constructor_00002() {
        assertEquals(HeartRateProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()), HeartRateProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()));
    }

    @Test
    public void test_getProfileName_00001() {
        assertEquals(HeartRateProfileBondedDatabaseHelper.PROFILE_NAME, HeartRateProfileBondedDatabaseHelper.getInstance(ApplicationProvider.getApplicationContext()).getProfileName());
    }
}
